requirements.txt
PyYAML >= 6.0.0, < 7.0.0
pydantic >= 2.9.0, < 3.0.0
python >= 3.12.0, < 4.0.0
requests >= 2.32.0, < 3.0.0
user-agents >= 2.2.0 < 3.0.0
